OpenConnectionRequest1: Protocol version is a byte, not an int

+++ src/main/kotlin/raknet/message/protocol/OpenConnectionRequest1.kt
@@ -8,7 +8,7 @@ import raknet.message.MessageType
 
 class OpenConnectionRequest1(
     var magic: Magic,
-    var protocolVersion: Int,
+    var protocolVersion: Byte,
     var mtuSize: Short
 ): OfflineMessage(MessageType.OPEN_CONNECTION_REQUEST_1.id()) {
 
@@ -17,7 +17,7 @@ class OpenConnectionRequest1(
     companion object {
         fun from(buffer: ByteBuf) = OpenConnectionRequest1(
             buffer.readMagic(),
-            buffer.readInt(),
+            buffer.readByte(),
             buffer.readableBytes().toShort()
         )
     }
